Upgrade OnDemandTV page to Typescript
