+++ src/app/pages/OnDemandTvPage/OnDemandTvPage.tsx
@@ -1,18 +1,7 @@
+/** @jsx jsx */
+import { jsx } from '@emotion/react';
 import React, { useContext } from 'react';
-import styled from '@emotion/styled';
-import {
-  GEL_SPACING,
-  GEL_SPACING_DBL,
-  GEL_SPACING_QUAD,
-  GEL_SPACING_QUIN,
-} from '#psammead/gel-foundations/src/spacings';
 import pathOr from 'ramda/src/pathOr';
-import {
-  GEL_GROUP_1_SCREEN_WIDTH_MAX,
-  GEL_GROUP_2_SCREEN_WIDTH_MAX,
-  GEL_GROUP_3_SCREEN_WIDTH_MAX,
-  GEL_GROUP_4_SCREEN_WIDTH_MIN,
-} from '#psammead/gel-foundations/src/breakpoints';
 import { formatUnixTimestamp } from '#psammead/psammead-timestamp-container/src/utilities';
 import ComscoreAnalytics from '#containers/ComscoreAnalytics';
 import Grid, { GelPageGrid } from '#components/Grid';
@@ -32,12 +21,16 @@ import { ServiceContext } from '../../contexts/ServiceContext';
 import MetadataContainer from '../../components/Metadata';
 import getPlaceholderImageUrl from '../../routes/utils/getPlaceholderImageUrl';
 import VisuallyHiddenText from '../../components/VisuallyHiddenText';
+import styles from './index.styles';
 
-const Wrapper = styled.div`
-  background-color: ${({ theme }) => theme.palette.GREY_10};
-`;
-
-const getGroups = (zero, one, two, three, four, five) => ({
+const getGroups = (
+  zero: number | boolean,
+  one: number | boolean,
+  two: number | boolean,
+  three: number | boolean,
+  four: number | boolean,
+  five: number | boolean,
+) => ({
   group0: zero,
   group1: one,
   group2: two,
@@ -46,29 +39,33 @@ const getGroups = (zero, one, two, three, four, five) => ({
   group5: five,
 });
 
-const StyledGelPageGrid = styled(GelPageGrid)`
-  padding-bottom: ${GEL_SPACING_QUAD};
-  @media (min-width: ${GEL_GROUP_4_SCREEN_WIDTH_MIN}) {
-    width: 100%;
-  }
-`;
+export interface OnDemandTVProps {
+  pageData: {
+    language: string;
+    headline: string;
+    shortSynopsis: string;
+    brandTitle: string;
+    releaseDateTimeStamp: string;
+    masterBrand?: string;
+    episodeId: string;
+    imageUrl: string;
+    promoBrandTitle: string;
+    thumbnailImageUrl: string;
+    durationISO8601: string;
+    recentEpisodes: string[];
+    episodeTitle: string;
+    mediumSynopsis: string;
+    contentType: 'player-episode';
+  };
+  mediaIsAvailable?: boolean;
+  MediaError?: React.ElementType;
+}
 
-const StyledVideoPlayer = styled(AVPlayer)`
-  @media (min-width: ${GEL_GROUP_4_SCREEN_WIDTH_MIN}) {
-    margin: ${GEL_SPACING_QUIN} 0 0;
-  }
-  @media (max-width: ${GEL_GROUP_3_SCREEN_WIDTH_MAX}) {
-    margin-top: ${GEL_SPACING_DBL};
-  }
-  @media (max-width: ${GEL_GROUP_2_SCREEN_WIDTH_MAX}) {
-    margin: ${GEL_SPACING_DBL} -${GEL_SPACING_DBL} 0;
-  }
-  @media (max-width: ${GEL_GROUP_1_SCREEN_WIDTH_MAX}) {
-    margin: ${GEL_SPACING} -${GEL_SPACING} 0;
-  }
-`;
-
-const OnDemandTvPage = ({ pageData, mediaIsAvailable, MediaError }) => {
+const OnDemandTvPage = ({
+  pageData,
+  mediaIsAvailable,
+  MediaError,
+}: OnDemandTVProps) => {
   const {
     language,
     headline,
@@ -84,6 +81,7 @@ const OnDemandTvPage = ({ pageData, mediaIsAvailable, MediaError }) => {
     recentEpisodes,
     episodeTitle,
     mediumSynopsis,
+    contentType,
   } = pageData;
 
   const { lang, timezone, datetimeLocale, service, translations, brandName } =
@@ -118,11 +116,11 @@ const OnDemandTvPage = ({ pageData, mediaIsAvailable, MediaError }) => {
     : headline;
 
   return (
-    <Wrapper>
+    <div css={styles.wrapper}>
       <ChartbeatAnalytics
         mediaPageType="TV"
         title={headline}
-        contentType={pageData?.contentType}
+        contentType={contentType}
       />
       <ATIAnalytics data={pageData} />
       <ComscoreAnalytics />
@@ -152,24 +150,28 @@ const OnDemandTvPage = ({ pageData, mediaIsAvailable, MediaError }) => {
             : []
         }
       />
-      <StyledGelPageGrid
+      {/* @ts-expect-error: Legacy grid expects `children` to be passed as props. However, due to coding best practices, we must nest children between the opening and closing tags */}
+      <GelPageGrid
         as="main"
         role="main"
         columns={getGroups(6, 6, 6, 6, 8, 20)}
         enableGelGutters
+        css={styles.pageGrid}
       >
+        {/* @ts-expect-error: Legacy grid expects `children` to be passed as props. However, due to coding best practices, we must nest children between the opening and closing tags */}
         <Grid
           item
           startOffset={getGroups(1, 1, 1, 1, 2, 5)}
           columns={getGroups(6, 6, 6, 6, 6, 12)}
           margins={getGroups(true, true, true, true, false, false)}
         >
-          <VisuallyHiddenText as="h1" tabIndex="-1" id="content">
+          {/* @ts-expect-error: Allow children to be passed as props */}
+          <VisuallyHiddenText as="h1" tabIndex={-1} id="content">
             {/* these must be concatenated for screen reader UX - #7062 */}
             {`${brandTitle}, ${formattedTimestamp}`}
           </VisuallyHiddenText>
           {mediaIsAvailable ? (
-            <StyledVideoPlayer
+            <AVPlayer
               embedUrl={embedUrl}
               assetId={episodeId}
               placeholderSrc={getPlaceholderImageUrl(imageUrl)}
@@ -180,6 +182,7 @@ const OnDemandTvPage = ({ pageData, mediaIsAvailable, MediaError }) => {
               skin="classic"
               showLoadingImage
               darkPlaceholder
+              css={styles.mediaPlayer}
             />
           ) : (
             <MediaError skin="video" />
@@ -192,6 +195,7 @@ const OnDemandTvPage = ({ pageData, mediaIsAvailable, MediaError }) => {
             ariaHidden
           />
         </Grid>
+        {/* @ts-expect-error: Legacy grid expects `children` to be passed as props. However, due to coding best practices, we must nest children between the opening and closing tags */}
         <Grid
           item
           columns={getGroups(6, 6, 6, 6, 5, 10)}
@@ -205,13 +209,16 @@ const OnDemandTvPage = ({ pageData, mediaIsAvailable, MediaError }) => {
             <FooterTimestamp releaseDateTimeStamp={releaseDateTimeStamp} />
           )}
         </Grid>
-      </StyledGelPageGrid>
+      </GelPageGrid>
 
       {hasRecentEpisodes && (
-        <StyledGelPageGrid
+        // @ts-expect-error: Legacy grid expects `children` to be passed as props. However, due to coding best practices, we must nest children between the opening and closing tags
+        <GelPageGrid
           columns={getGroups(6, 6, 6, 6, 8, 20)}
           enableGelGutters
+          css={styles.pageGrid}
         >
+          {/* @ts-expect-error: Legacy grid expects `children` to be passed as props. However, due to coding best practices, we must nest children between the opening and closing tags */}
           <Grid
             item
             startOffset={getGroups(1, 1, 1, 1, 2, 5)}
@@ -223,9 +230,9 @@ const OnDemandTvPage = ({ pageData, mediaIsAvailable, MediaError }) => {
               episodes={recentEpisodes}
             />
           </Grid>
-        </StyledGelPageGrid>
+        </GelPageGrid>
       )}
-    </Wrapper>
+    </div>
   );
 };
